Setup Error Handling on the scale of the server to handle production and development errors with the additon to handle not found routes

+++ server/src/modules/user/services/user.service.ts
@@ -2,6 +2,8 @@ import { IUser } from '../Interfaces/IUser';
 import { generateUserId } from '../../../utils/user-data-handler';
 import { hash } from '../../../utils/hashing-handler';
 import UserRepository from '../repositories/user.repository';
+import { NextFunction } from 'express';
+import AppError from '../../../utils/app-error';
 
 const userRepository = new UserRepository();
 
@@ -23,12 +25,10 @@ class UserService {
   /*
    * Update User Service
    */
-  static async updateUser(updatedData: any) {
+  static async updateUser(updatedData: any, next: NextFunction) {
     const { userId, ...newData } = updatedData;
-    const user = await this.getUser(userId);
-    if (!user) {
-      // TODO : throw an error
-    }
+    const user = await this.getUser(userId, next);
+    if(!user) return;
     const userData: IUser = {
       ...user,
       ...newData,
@@ -38,7 +38,9 @@ class UserService {
   /*
    * Delete User Service
    */
-  static async deleteUser(userId: string) {
+  static async deleteUser(userId: string, next: NextFunction) {
+    const user = await this.getUser(userId, next);
+    if(!user) return;
     return userRepository.deleteUser(userId);
   }
   /*
@@ -50,8 +52,19 @@ class UserService {
   /*
    * Get user Service
    */
-  static async getUser(userId: string) {
-    return userRepository.getUser(userId);
+  static async getUser(userId: string, next: NextFunction) {
+    const user = await userRepository.getUser(userId);
+    if (!user) {
+       next(
+        new AppError(
+          404,
+          `User with ID : ${userId} is not found, use a valid ID and try again!`
+        )
+        
+      );
+      return;
+    }
+    return user;
   }
 }
 export default UserService;
